#72 updated service methods, typos

+++ source/Almostengr.VideoProcessor.Core/TechTalk/TechTalkService.cs
@@ -65,133 +65,6 @@ public sealed class TechTalkService : BaseVideoService, ITechTalkVideoService, I
         }
     }
 
-    internal async Task ProcessIncomingVideoProjectAsync(CancellationToken cancellationToken)
-    {
-        TechTalkVideoFile? archiveFile = null;
-
-        try
-        {
-            string? selectedTarballFilePath = _fileSystemService.GetRandomFileByExtensionFromDirectory(
-                IncomingDirectory, FileExtension.Tar);
-
-            if (string.IsNullOrEmpty(selectedTarballFilePath))
-            {
-                return;
-            }
-
-            archiveFile = new TechTalkVideoFile(new VideoProjectArchiveFile(selectedTarballFilePath));
-
-            _fileSystemService.DeleteDirectory(WorkingDirectory);
-            _fileSystemService.CreateDirectory(WorkingDirectory);
-
-            await _tarballService.ExtractTarballContentsAsync(archiveFile.FilePath, WorkingDirectory, cancellationToken);
-
-            StopProcessingIfKdenliveFileExists(WorkingDirectory);
-            StopProcessingIfDetailsTxtFileExists(WorkingDirectory);
-            StopProcessingIfFfmpegInputTxtFileExists(WorkingDirectory);
-
-            _fileSystemService.PrepareAllFilesInDirectory(WorkingDirectory);
-
-            // todo normailze audio files
-
-            var audioFiles = _fileSystemService.GetFilesInDirectory(WorkingDirectory)
-                .Where(f => f.EndsWith(FileExtension.Mp3.Value, StringComparison.OrdinalIgnoreCase))
-                .Select(f => new AudioFile(f))
-                .ToList();
-
-            foreach (var audioFile in audioFiles)
-            {
-                var video = _fileSystemService.GetFilesInDirectory(WorkingDirectory)
-                    .Where(f => f.StartsWith(audioFile.FilePath.Replace(FileExtension.Mp3.Value, string.Empty)) && !f.Equals(audioFile.FilePath))
-                    .Select(f => new TechTalkVideoFile(f))
-                    .Single();
-
-                video.SetAudioFile(audioFile);
-
-                string tsOutputFilePath = Path.Combine(
-                    WorkingDirectory,
-                    Path.GetFileNameWithoutExtension(video.FileName()) + FileExtension.Ts.Value);
-
-                await _ffmpegService.AddAccAudioToVideoAsync(
-                    video.FilePath, video.AudioFilePath(), tsOutputFilePath, cancellationToken);
-
-                _fileSystemService.DeleteFile(video.FilePath);
-            }
-
-            var videoSegments = _fileSystemService.GetFilesInDirectory(WorkingDirectory)
-                .Where(f => f.EndsWith(FileExtension.Mp4.Value, StringComparison.OrdinalIgnoreCase) ||
-                    f.EndsWith(FileExtension.Mkv.Value, StringComparison.OrdinalIgnoreCase) ||
-                    f.EndsWith(FileExtension.Mov.Value, StringComparison.OrdinalIgnoreCase))
-                .Select(f => new TechTalkVideoFile(f))
-                .ToList();
-
-            foreach (var video in videoSegments)
-            {
-                var result = await _ffmpegService.FfprobeAsync($"\"{video.FilePath}\"", WorkingDirectory, cancellationToken);
-
-                if (result.stdErr.ToLower().Contains(Constant.Audio))
-                {
-                    await _ffmpegService.ConvertMp4VideoFileToTsFormatAsync(
-                        video.FilePath,
-                        video.FilePath.Replace(FileExtension.Mp4.Value, FileExtension.Ts.Value)
-                            .Replace(FileExtension.Mkv.Value, FileExtension.Ts.Value)
-                            .Replace(FileExtension.Mov.Value, FileExtension.Ts.Value),
-                        cancellationToken);
-                    continue;
-                }
-
-                video.SetAudioFile(_musicService.GetRandomMixTrack());
-
-                string tsOutputFilePath = Path.Combine(WorkingDirectory, video.TsOutputFileName());
-
-                await _ffmpegService.AddAccAudioToVideoAsync(
-                    video.FilePath, video.AudioFilePath(), tsOutputFilePath, cancellationToken);
-
-                _fileSystemService.DeleteFile(video.FilePath);
-            }
-
-            string? ffmpegInputFilePath = _fileSystemService.GetFilesInDirectory(WorkingDirectory)
-                .Where(f => f.EndsWith(FileExtension.FfmpegInput.Value))
-                .SingleOrDefault();
-
-            if (string.IsNullOrEmpty(ffmpegInputFilePath))
-            {
-                var tsVideoFiles = _fileSystemService.GetFilesInDirectory(WorkingDirectory)
-                    .Where(f => f.EndsWith(FileExtension.Ts.Value, StringComparison.OrdinalIgnoreCase))
-                    .OrderBy(f => f);
-
-                ffmpegInputFilePath = Path.Combine(WorkingDirectory, Constant.FfmpegInputFileName);
-                CreateFfmpegInputFile(tsVideoFiles.ToArray(), ffmpegInputFilePath);
-            }
-
-            string outputVideoFilePath = Path.Combine(WorkingDirectory, archiveFile.OutputFileName());
-
-            await _ffmpegService.RenderVideoWithInputFileAndFiltersAsync(
-                ffmpegInputFilePath, archiveFile.VideoFilters(), outputVideoFilePath, cancellationToken);
-
-            _fileSystemService.SaveFileContents(
-                Path.Combine(IncomingDirectory, archiveFile.ThumbnailFileName()), archiveFile.Title());
-
-            _fileSystemService.MoveFile(archiveFile.FilePath, Path.Combine(ArchiveDirectory, archiveFile.FileName()));
-            _fileSystemService.MoveFile(
-                outputVideoFilePath, Path.Combine(UploadingDirectory, archiveFile.OutputFileName()));
-
-            _fileSystemService.DeleteDirectory(WorkingDirectory);
-        }
-        catch (Exception ex)
-        {
-            _loggerService.LogError(ex, ex.Message);
-
-            if (archiveFile != null)
-            {
-                _loggerService.LogErrorProcessingFile(archiveFile.FilePath, ex);
-                _fileSystemService.MoveFile(archiveFile.FilePath, archiveFile.FilePath + FileExtension.Err.Value);
-            }
-
-            _fileSystemService.DeleteDirectory(WorkingDirectory);
-        }
-    }
-
     public override async Task ProcessIncomingTarballFilesAsync(CancellationToken cancellationToken)
     {
         TechTalkVideoFile? archiveFile = null;
@@ -228,7 +101,7 @@ public sealed class TechTalkService : BaseVideoService, ITechTalkVideoService, I
             foreach (var audioFile in audioFiles)
             {
                 var video = _fileSystemService.GetFilesInDirectory(WorkingDirectory)
-                    .Where(f => f.StartsWith(audioFile.FilePath.Replace(FileExtension.Mp3.Value, string.Empty)) && !f.Equals(audioFile.FilePath))
+                    .Where(f => f.StartsWith(audioFile.FilePath.Replace(FileExtension.Mp3.Value, string.Empty, StringComparison.OrdinalIgnoreCase)) && !f.Equals(audioFile.FilePath))
                     .Select(f => new TechTalkVideoFile(f))
                     .Single();
 
@@ -244,12 +117,11 @@ public sealed class TechTalkService : BaseVideoService, ITechTalkVideoService, I
                 _fileSystemService.DeleteFile(video.FilePath);
             }
 
-            var mp4MkvVideoFiles = _fileSystemService.GetFilesInDirectory(WorkingDirectory)
-                .Where(f => f.EndsWith(FileExtension.Mp4.Value, StringComparison.OrdinalIgnoreCase) || f.EndsWith(FileExtension.Mkv.Value, StringComparison.OrdinalIgnoreCase))
+            var videoSegment = GetVideoFilesInDirectory(WorkingDirectory)
                 .Select(f => new TechTalkVideoFile(f))
                 .ToList();
 
-            foreach (var video in mp4MkvVideoFiles)
+            foreach (var video in videoSegment)
             {
                 var result = await _ffmpegService.FfprobeAsync($"\"{video.FilePath}\"", WorkingDirectory, cancellationToken);
 
@@ -257,7 +129,8 @@ public sealed class TechTalkService : BaseVideoService, ITechTalkVideoService, I
                 {
                     await _ffmpegService.ConvertMp4VideoFileToTsFormatAsync(
                         video.FilePath,
-                        video.FilePath.Replace(FileExtension.Mp4.Value, FileExtension.Ts.Value).Replace(FileExtension.Mkv.Value, FileExtension.Ts.Value),
+                        video.FilePath.Replace(FileExtension.Mp4.Value, FileExtension.Ts.Value)
+                            .Replace(FileExtension.Mkv.Value, FileExtension.Ts.Value),
                         cancellationToken);
                     continue;
                 }
@@ -282,7 +155,7 @@ public sealed class TechTalkService : BaseVideoService, ITechTalkVideoService, I
                     .Where(f => f.EndsWith(FileExtension.Ts.Value, StringComparison.OrdinalIgnoreCase))
                     .OrderBy(f => f);
 
-                ffmpegInputFilePath = Path.Combine(WorkingDirectory, "videos" + FileExtension.FfmpegInput.Value);
+                ffmpegInputFilePath = Path.Combine(WorkingDirectory, Constant.FfmpegInputFileName);
                 CreateFfmpegInputFile(tsVideoFiles.ToArray(), ffmpegInputFilePath);
             }
 
@@ -347,8 +220,15 @@ public sealed class TechTalkService : BaseVideoService, ITechTalkVideoService, I
 
         try
         {
-            srtFile = new SrtSubtitleFile(_fileSystemService.GetRandomFileByExtensionFromDirectory(
-                WorkingDirectory, FileExtension.Srt));
+            string? randomFile = _fileSystemService.GetRandomFileByExtensionFromDirectory(
+                WorkingDirectory, FileExtension.Srt);
+
+            if (string.IsNullOrEmpty(randomFile))
+            {
+                return;
+            }
+
+            srtFile = new SrtSubtitleFile(randomFile);
 
             _fileSystemService.DeleteDirectory(WorkingDirectory);
             _fileSystemService.CreateDirectory(WorkingDirectory);
